Rename the first version of the wizard.
